Add parser
Now the code tries to find a match within an input string.
To avoid complications with calculating true length of pattern
(since "[abcde]" is actually a one symbol) the code tries to match
pattern to input_line, input_line[1:], input_line[2:] and so on.

Also fixed bugs in negative groups.