Add group splitting function
It would be much easier to integrate this function into exsiting code
if the function would just split the group based on "|" and their depth.

For example "((a|b)|(b|c))" would be splitted as "(a|b)" and "(b|c)".
And then "parse" function would take care of the rest.