FINALLY IT UNDERSTANDS GROUPS!!!!!
The code was too complicated. I thought that "+" case should be similar
for every pattern. Thus ALTERATION case should process only one
"iteration" and "+" case should repeat.

There was a problem that "+" haven't repeated to much; for instance,
cases like "aaaaaa" "a$" might have not evaluated properly.
Gemini have noticed this.
Now the code loops the pattern as much as it can and then tries to
evaluate each possible branch.

Also shifted "?" in the beggining of the function to properly evaluate,
say, the pattern "a1?" on string "a".