Add support for backreferrences. Refactor code
I was thinking that nested backreferrences should be
valid only within the group, so I've changed the code a little
so that each group is itself a different matching task.
But after considering nested backreferrence just a litlle more
carefully I realized that we should have global list or hash-map
for backreferrences.